+++ R/emis_evap.R
@@ -16,8 +16,6 @@
 #' for vehicles with fuel injection and returnless fuel systems.
 #'  If x has units 'lkm', the units of ed must be 'g/km',
 #'  other case, this are simply g/trip or g/proced
-#' @param carb fraction of gasoline vehicles with carburator or fuel return system.
-#' @param p Fraction of trips finished with hot engine
 #' @param hotc average running losses or soak evaporative factor for vehicles with
 #' carburator or fuel return system
 #' for vehicles with fuel injection and returnless fuel systems.
@@ -26,7 +24,11 @@
 #' for vehicles with carburator or fuel return system
 #' for vehicles with fuel injection and returnless fuel systems.
 #'  If x has units 'lkm', the units of ed must be 'g/km',
+#' @param carb fraction of gasoline vehicles with carburator or fuel return system.
+#' @param p Fraction of trips finished with hot engine
+#' @param params Character; Add columns with information to returning data.frame
 #' @param pro_month Numeric; montly profile to distribuite annual mileage in each month.
+#' @param verbose Logical; To show more information
 #' @return numeric vector of emission estimation in grams
 #' @importFrom units as_units
 #' @references Mellios G and Ntziachristos 2016. Gasoline evaporation. In:
@@ -44,13 +46,22 @@ emis_evap <- function(veh,
                       ed,
                       hotfi, hotc, warmc,
                       carb = 0, p,
-                      pro_month) { # hot or warm soak
+                      params,
+                      pro_month,
+                      verbose = FALSE) { # hot or warm soak
+  # If months exists, it must the last column to be added
   # Check y
   if(!missing(x)){
     if(units(x)$numerator == "km"){
       message('Emission factors must have units g/km')
     }
   }
+  # Checking sf
+  if(any(class(veh) %in% "sf")){
+    if(verbose) message("converting sf to data.frame")
+    veh <- sf::st_set_geometry(veh, NULL)
+  }
+
   # Check x for data.frame
   if(is.data.frame(veh) ){
     # DO I need check for 'Vehicles'?
@@ -63,7 +74,9 @@ emis_evap <- function(veh,
   } else {
     veh <- as.numeric(veh)
   }
-
+ if(!missing(params)){
+  if(verbose) message("Params will be added")
+ }
 
 
   # ed
@@ -80,18 +93,42 @@ emis_evap <- function(veh,
             veh[, i]*x[i]*ed[[j]][, i]*pro_month[j]
           }))
           e <- Emissions(e)
+          if(!missing(params)){
+            for (i in 1:length(params)){
+              e <- as.data.frame(e)
+              e[, params[i]] <- params[i]
+            }
+          }
+
           e$month <- mes[j]
           e
         }))
+        if(verbose) cat("Sum of emissions:", sum(e[-"month"]), "\n")
+
       } else {
         if(!is.data.frame(ed)) stop("'ed' must be a data.frame with the same dimensions of 'veh'")
         e <- Emissions(do.call("cbind", lapply(1:ncol(veh), function(i){
           veh[, i]*x[i]*ed[, i]
         })))
+        if(verbose) cat("Sum of emissions:", sum(e), "\n")
+        if(!missing(params)){
+          for (i in 1:length(params)){
+            e <- as.data.frame(e)
+            e[, params[i]] <- params[i]
+          }
+        }
 
       }
     } else {
       e <- Emissions(veh*x*ed)
+      if(verbose) cat("Sum of emissions:", sum(e), "\n")
+      if(!missing(params)){
+        for (i in 1:length(params)){
+          e <- as.data.frame(e)
+          e[, params[i]] <- params[i]
+        }
+      }
+
     }
 
   } else {
@@ -106,9 +143,17 @@ emis_evap <- function(veh,
               veh[, i]*x[i]*(carb*(p*hotc[[j]][, i]+(1-p)*warmc[[j]][, i])+(1-carb)*hotfi[[j]][, i])*pro_month[j]
             }))
             e <- Emissions(e)
+            if(!missing(params)){
+              for (i in 1:length(params)){
+                e <- as.data.frame(e)
+                e[, params[i]] <- params[i]
+              }
+            }
+
             e$month <- mes[j]
             e
           }))
+          if(verbose) cat("Sum of emissions:", sum(e[-"month"]), "\n")
 
         } else {
           if(!is.data.frame(hotc)) stop("hotc' must be a data.frame with the same dimensions of 'veh'")
@@ -117,11 +162,25 @@ emis_evap <- function(veh,
           e <- Emissions(do.call("cbind",lapply(1:ncol(veh), function(i){
             veh[, i]*x[i]*(carb*(p*hotc[, i]+(1-p)*warmc[, i])+(1-carb)*hotfi[, i])
           })))
+          if(verbose) cat("Sum of emissions:", sum(e), "\n")
+          if(!missing(params)){
+            for (i in 1:length(params)){
+              e <- as.data.frame(e)
+              e[, params[i]] <- params[i]
+            }
+          }
 
         }
 
       } else {
         e <- veh*x*(carb*(p*hotc+(1-p)*warmc)+(1-carb)*hotfi)
+        if(!missing(params)){
+          for (i in 1:length(params)){
+            e <- as.data.frame(e)
+            e[, params[i]] <- params[i]
+          }
+        }
+
       }
     } else if (carb < 0){
       stop("carb is a positive fraction or 0")
@@ -136,22 +195,43 @@ emis_evap <- function(veh,
               veh[, i]*x[i]*hotfi[[j]][, i]*pro_month[j]
             }))
             e <- Emissions(e)
+            if(!missing(params)){
+              for (i in 1:length(params)){
+                e <- as.data.frame(e)
+                e[, params[i]] <- params[i]
+              }
+            }
             e$month <- mes[j]
             e
           }))
+          if(verbose) cat("Sum of emissions:", sum(e[-"month"]), "\n")
 
         } else {
           if(!is.data.frame(hotfi)) stop("'hotfi' must be a data.frame with the same dimensions of 'veh'")
           e <- Emissions(do.call("cbind",lapply(1:ncol(veh), function(i){
             veh[, i]*x[i]*hotfi[, i]
           })))
+          if (verbose) cat("Sum of emissions:", sum(e), "\n")
+          if(!missing(params)){
+            for (i in 1:length(params)){
+              e <- as.data.frame(e)
+              e[, params[i]] <- params[i]
+            }
+          }
 
         }
       } else {
         e <- veh*x*hotfi
+        if(!missing(params)){
+          for (i in 1:length(params)){
+            e <- as.data.frame(e)
+            e[, params[i]] <- params[i]
+          }
+        }
+
       }
+
     }
   }
-
   return(e)
 }